requirements.txt
streamlit
streamlit-qr-scanner
pdfplumber
requests
gspread
google-auth